Attempting to fix datepicker bug
If you load the page, add a bunch of empty projects, and then start working with the dates, all the datepickers work fine. If you load the page, fill in the first project, and THEN add another the datepickers in the new one won't open. I think this should fix it.