Fixed slightly different datepicker bug
While testing out example 4 (which has a lot of duplicate days) I noticed that when you add a new project, while the dates show as populated by the previous project, if you don't actually go through the datepicker and pick a date anyway that date won't properly register. When you go to calculate it's in there as a null and the calculator throws an error. I think it's because I was initializing the datepickers when the text input gets focus (to handle new ones being added on the fly). So no focus, no datepicker, nothing to get the date from. I thought about blanking out the dates on the new projects, but it's handy to have them there because when you open the datepicker it will automatically open to the correct month. So instead I'm forcing initialization of the new datepicker fields as soon as the new project html gets added.