Fixed bug with empty dates
Realized if I load the page and submit the form blank I still get a reimbursement. Validation was not kicking out dates that were present but not filled in with a value, and javascript's Date was just defaulting them to the epoch. Disallowing null/empty/undefined dates and adding another couple of tests to the suite to cover it.