Fixed key conversion error
It looks like I was having a key conversion problem. I'm using the getTime() value of each day as the key in the map, and that's an integer. But it looks like when I was iterating through all keys in the map it was giving them to me as strings ("4" instead of 4). Passing the string version to the date constructor was giving me invalid dates - I guess seeing the string it was trying to parse it instead of just accepting it as the time value. So "yesterday" and "tomorrow" were always invalid dates, thus not found in the map, thus the cost function was interpreting every day as a travel day. Hopefully that fixes the incorrect totals Al reported, although I still don't know how I messed them up the same way when checking them all by hand.